requirements.txt
asgiref==3.5.2
black==22.3.0
click==8.1.3
Django==4.0.5
django-cors-headers==3.13.0
django-environ==0.9.0
django-resized==1.0.0
djangorestframework==3.13.1
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.1.1
platformdirs==2.5.2
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.2.0
